requirements.txt
# Core scientific computing
numpy>=1.24.0
scipy>=1.11.0
pandas>=2.0.0

# Machine learning
scikit-learn>=1.3.0
xgboost>=2.0.0

# Explainability
shap>=0.43.0

# Visualization
matplotlib>=3.7.0
seaborn>=0.13.0

# Configuration
pyyaml>=6.0

# Jupyter notebooks
jupyter>=1.0.0
ipykernel>=6.25.0

# Testing
pytest>=7.4.0

# Utilities
tqdm>=4.66.0
